refactor: extract post dtos
