+++ backend/src/main/java/com/openisle/dto/CategoryDto.java
@@ -1,0 +1,16 @@
+package com.openisle.dto;
+
+import lombok.Data;
+
+/**
+ * DTO representing a post category.
+ */
+@Data
+public class CategoryDto {
+    private Long id;
+    private String name;
+    private String description;
+    private String icon;
+    private String smallIcon;
+}
+
