+++ backend/src/main/java/com/openisle/dto/PostRequest.java
@@ -1,0 +1,18 @@
+package com.openisle.dto;
+
+import lombok.Data;
+
+import java.util.List;
+
+/**
+ * Request body for creating or updating a post.
+ */
+@Data
+public class PostRequest {
+    private Long categoryId;
+    private String title;
+    private String content;
+    private List<Long> tagIds;
+    private String captcha;
+}
+
