+++ backend/src/main/java/com/openisle/dto/TagDto.java
@@ -1,0 +1,16 @@
+package com.openisle.dto;
+
+import lombok.Data;
+
+/**
+ * DTO representing a tag.
+ */
+@Data
+public class TagDto {
+    private Long id;
+    private String name;
+    private String description;
+    private String icon;
+    private String smallIcon;
+}
+
